+++ BackendCConecta/BackendCConecta/Infraestructura/Repositorios/Ubicaciones/LugaresReferenciaRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace BackendCConecta.Infraestructura.Repositorios.Ubicaciones
+{
+    public class LugaresReferenciaRepository
+    {
+    }
+}
