--- BackendCConecta/BackendCConecta/Infraestructura/Repositorios/Ubicationes/LugaresReferenciaRepository.cs
@@ -1,6 +1,0 @@
-﻿namespace BackendCConecta.Infraestructura.Repositorios.LugaresReferencia
-{
-    public class LugaresReferenciaRepository
-    {
-    }
-}
